Move script to flash
