+++ src/main.cpp
@@ -18,7 +18,7 @@
 #include "./apps/tools/print_debug.h"
 #include "./apps/tools/time_from_web.h"
 #include "./apps/tools/water_level.h"
-#include "./apps/main/luapp.h"
+#include "./apps/main/luaapp.h"
 #include "./overlays/overlays.h"
 #include "apps/lua/mylua_example.h"
 #if defined(GPS_EDITION)
